+++ src/Courier.Tests/Models/Notifications/NotificationListResponseTest.cs
@@ -16,15 +16,15 @@ public class NotificationListResponseTest : TestBase
             Paging = new() { More = true, Cursor = "cursor" },
             Results =
             [
-                new ResultNotification()
+                new Notification()
                 {
                     ID = "id",
                     CreatedAt = 0,
                     EventIds = ["string"],
-                    Note = "note",
                     Routing = new() { Channels = ["string"], Method = Method.All },
                     TopicID = "topic_id",
                     UpdatedAt = 0,
+                    Note = "note",
                     Tags = new([new() { ID = "id", Name = "name" }]),
                     Title = "title",
                 },
@@ -34,15 +34,15 @@ public class NotificationListResponseTest : TestBase
         Paging expectedPaging = new() { More = true, Cursor = "cursor" };
         List<Result> expectedResults =
         [
-            new ResultNotification()
+            new Notification()
             {
                 ID = "id",
                 CreatedAt = 0,
                 EventIds = ["string"],
-                Note = "note",
                 Routing = new() { Channels = ["string"], Method = Method.All },
                 TopicID = "topic_id",
                 UpdatedAt = 0,
+                Note = "note",
                 Tags = new([new() { ID = "id", Name = "name" }]),
                 Title = "title",
             },
@@ -64,15 +64,15 @@ public class NotificationListResponseTest : TestBase
             Paging = new() { More = true, Cursor = "cursor" },
             Results =
             [
-                new ResultNotification()
+                new Notification()
                 {
                     ID = "id",
                     CreatedAt = 0,
                     EventIds = ["string"],
-                    Note = "note",
                     Routing = new() { Channels = ["string"], Method = Method.All },
                     TopicID = "topic_id",
                     UpdatedAt = 0,
+                    Note = "note",
                     Tags = new([new() { ID = "id", Name = "name" }]),
                     Title = "title",
                 },
@@ -96,15 +96,15 @@ public class NotificationListResponseTest : TestBase
             Paging = new() { More = true, Cursor = "cursor" },
             Results =
             [
-                new ResultNotification()
+                new Notification()
                 {
                     ID = "id",
                     CreatedAt = 0,
                     EventIds = ["string"],
-                    Note = "note",
                     Routing = new() { Channels = ["string"], Method = Method.All },
                     TopicID = "topic_id",
                     UpdatedAt = 0,
+                    Note = "note",
                     Tags = new([new() { ID = "id", Name = "name" }]),
                     Title = "title",
                 },
@@ -121,15 +121,15 @@ public class NotificationListResponseTest : TestBase
         Paging expectedPaging = new() { More = true, Cursor = "cursor" };
         List<Result> expectedResults =
         [
-            new ResultNotification()
+            new Notification()
             {
                 ID = "id",
                 CreatedAt = 0,
                 EventIds = ["string"],
-                Note = "note",
                 Routing = new() { Channels = ["string"], Method = Method.All },
                 TopicID = "topic_id",
                 UpdatedAt = 0,
+                Note = "note",
                 Tags = new([new() { ID = "id", Name = "name" }]),
                 Title = "title",
             },
@@ -151,15 +151,15 @@ public class NotificationListResponseTest : TestBase
             Paging = new() { More = true, Cursor = "cursor" },
             Results =
             [
-                new ResultNotification()
+                new Notification()
                 {
                     ID = "id",
                     CreatedAt = 0,
                     EventIds = ["string"],
-                    Note = "note",
                     Routing = new() { Channels = ["string"], Method = Method.All },
                     TopicID = "topic_id",
                     UpdatedAt = 0,
+                    Note = "note",
                     Tags = new([new() { ID = "id", Name = "name" }]),
                     Title = "title",
                 },
@@ -177,15 +177,15 @@ public class NotificationListResponseTest : TestBase
             Paging = new() { More = true, Cursor = "cursor" },
             Results =
             [
-                new ResultNotification()
+                new Notification()
                 {
                     ID = "id",
                     CreatedAt = 0,
                     EventIds = ["string"],
-                    Note = "note",
                     Routing = new() { Channels = ["string"], Method = Method.All },
                     TopicID = "topic_id",
                     UpdatedAt = 0,
+                    Note = "note",
                     Tags = new([new() { ID = "id", Name = "name" }]),
                     Title = "title",
                 },
@@ -203,15 +203,15 @@ public class ResultTest : TestBase
     [Fact]
     public void NotificationValidationWorks()
     {
-        Result value = new ResultNotification()
+        Result value = new Notification()
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
             Tags = new([new() { ID = "id", Name = "name" }]),
             Title = "title",
         };
@@ -238,15 +238,15 @@ public class ResultTest : TestBase
     [Fact]
     public void NotificationSerializationRoundtripWorks()
     {
-        Result value = new ResultNotification()
+        Result value = new Notification()
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
             Tags = new([new() { ID = "id", Name = "name" }]),
             Title = "title",
         };
@@ -277,20 +277,20 @@ public class ResultTest : TestBase
     }
 }
 
-public class ResultNotificationTest : TestBase
+public class NotificationTest : TestBase
 {
     [Fact]
     public void FieldRoundtrip_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
             Tags = new([new() { ID = "id", Name = "name" }]),
             Title = "title",
         };
@@ -298,10 +298,10 @@ public class ResultNotificationTest : TestBase
         string expectedID = "id";
         long expectedCreatedAt = 0;
         List<string> expectedEventIds = ["string"];
-        string expectedNote = "note";
         MessageRouting expectedRouting = new() { Channels = ["string"], Method = Method.All };
         string expectedTopicID = "topic_id";
         long expectedUpdatedAt = 0;
+        string expectedNote = "note";
         Tags expectedTags = new([new() { ID = "id", Name = "name" }]);
         string expectedTitle = "title";
 
@@ -312,10 +312,10 @@ public class ResultNotificationTest : TestBase
         {
             Assert.Equal(expectedEventIds[i], model.EventIds[i]);
         }
-        Assert.Equal(expectedNote, model.Note);
         Assert.Equal(expectedRouting, model.Routing);
         Assert.Equal(expectedTopicID, model.TopicID);
         Assert.Equal(expectedUpdatedAt, model.UpdatedAt);
+        Assert.Equal(expectedNote, model.Note);
         Assert.Equal(expectedTags, model.Tags);
         Assert.Equal(expectedTitle, model.Title);
     }
@@ -323,21 +323,21 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void SerializationRoundtrip_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
             Tags = new([new() { ID = "id", Name = "name" }]),
             Title = "title",
         };
 
         string json = JsonSerializer.Serialize(model, ModelBase.SerializerOptions);
-        var deserialized = JsonSerializer.Deserialize<ResultNotification>(
+        var deserialized = JsonSerializer.Deserialize<Notification>(
             json,
             ModelBase.SerializerOptions
         );
@@ -348,21 +348,21 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void FieldRoundtripThroughSerialization_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
             Tags = new([new() { ID = "id", Name = "name" }]),
             Title = "title",
         };
 
         string element = JsonSerializer.Serialize(model, ModelBase.SerializerOptions);
-        var deserialized = JsonSerializer.Deserialize<ResultNotification>(
+        var deserialized = JsonSerializer.Deserialize<Notification>(
             element,
             ModelBase.SerializerOptions
         );
@@ -371,10 +371,10 @@ public class ResultNotificationTest : TestBase
         string expectedID = "id";
         long expectedCreatedAt = 0;
         List<string> expectedEventIds = ["string"];
-        string expectedNote = "note";
         MessageRouting expectedRouting = new() { Channels = ["string"], Method = Method.All };
         string expectedTopicID = "topic_id";
         long expectedUpdatedAt = 0;
+        string expectedNote = "note";
         Tags expectedTags = new([new() { ID = "id", Name = "name" }]);
         string expectedTitle = "title";
 
@@ -385,10 +385,10 @@ public class ResultNotificationTest : TestBase
         {
             Assert.Equal(expectedEventIds[i], deserialized.EventIds[i]);
         }
-        Assert.Equal(expectedNote, deserialized.Note);
         Assert.Equal(expectedRouting, deserialized.Routing);
         Assert.Equal(expectedTopicID, deserialized.TopicID);
         Assert.Equal(expectedUpdatedAt, deserialized.UpdatedAt);
+        Assert.Equal(expectedNote, deserialized.Note);
         Assert.Equal(expectedTags, deserialized.Tags);
         Assert.Equal(expectedTitle, deserialized.Title);
     }
@@ -396,12 +396,49 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void Validation_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
+            Routing = new() { Channels = ["string"], Method = Method.All },
+            TopicID = "topic_id",
+            UpdatedAt = 0,
             Note = "note",
+            Tags = new([new() { ID = "id", Name = "name" }]),
+            Title = "title",
+        };
+
+        model.Validate();
+    }
+
+    [Fact]
+    public void OptionalNonNullablePropertiesUnsetAreNotSet_Works()
+    {
+        var model = new Notification
+        {
+            ID = "id",
+            CreatedAt = 0,
+            EventIds = ["string"],
+            Routing = new() { Channels = ["string"], Method = Method.All },
+            TopicID = "topic_id",
+            UpdatedAt = 0,
+            Tags = new([new() { ID = "id", Name = "name" }]),
+            Title = "title",
+        };
+
+        Assert.Null(model.Note);
+        Assert.False(model.RawData.ContainsKey("note"));
+    }
+
+    [Fact]
+    public void OptionalNonNullablePropertiesUnsetValidation_Works()
+    {
+        var model = new Notification
+        {
+            ID = "id",
+            CreatedAt = 0,
+            EventIds = ["string"],
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
@@ -413,17 +450,60 @@ public class ResultNotificationTest : TestBase
     }
 
     [Fact]
-    public void OptionalNullablePropertiesUnsetAreNotSet_Works()
+    public void OptionalNonNullablePropertiesSetToNullAreNotSet_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Tags = new([new() { ID = "id", Name = "name" }]),
+            Title = "title",
+
+            // Null should be interpreted as omitted for these properties
+            Note = null,
+        };
+
+        Assert.Null(model.Note);
+        Assert.False(model.RawData.ContainsKey("note"));
+    }
+
+    [Fact]
+    public void OptionalNonNullablePropertiesSetToNullValidation_Works()
+    {
+        var model = new Notification
+        {
+            ID = "id",
+            CreatedAt = 0,
+            EventIds = ["string"],
+            Routing = new() { Channels = ["string"], Method = Method.All },
+            TopicID = "topic_id",
+            UpdatedAt = 0,
+            Tags = new([new() { ID = "id", Name = "name" }]),
+            Title = "title",
+
+            // Null should be interpreted as omitted for these properties
+            Note = null,
+        };
+
+        model.Validate();
+    }
+
+    [Fact]
+    public void OptionalNullablePropertiesUnsetAreNotSet_Works()
+    {
+        var model = new Notification
+        {
+            ID = "id",
+            CreatedAt = 0,
+            EventIds = ["string"],
+            Routing = new() { Channels = ["string"], Method = Method.All },
+            TopicID = "topic_id",
+            UpdatedAt = 0,
+            Note = "note",
         };
 
         Assert.Null(model.Tags);
@@ -435,15 +515,15 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void OptionalNullablePropertiesUnsetValidation_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
         };
 
         model.Validate();
@@ -452,15 +532,15 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void OptionalNullablePropertiesSetToNullAreSetToNull_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
 
             Tags = null,
             Title = null,
@@ -475,15 +555,15 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void OptionalNullablePropertiesSetToNullValidation_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
 
             Tags = null,
             Title = null,
@@ -495,20 +575,20 @@ public class ResultNotificationTest : TestBase
     [Fact]
     public void CopyConstructor_Works()
     {
-        var model = new ResultNotification
+        var model = new Notification
         {
             ID = "id",
             CreatedAt = 0,
             EventIds = ["string"],
-            Note = "note",
             Routing = new() { Channels = ["string"], Method = Method.All },
             TopicID = "topic_id",
             UpdatedAt = 0,
+            Note = "note",
             Tags = new([new() { ID = "id", Name = "name" }]),
             Title = "title",
         };
 
-        ResultNotification copied = new(model);
+        Notification copied = new(model);
 
         Assert.Equal(model, copied);
     }